fixed a bug where the configuration item type for the numeric type in .env was incorrectly typed

+++ publish/jwt.php
@@ -98,7 +98,7 @@ return [
     |
     */
 
-    'ttl' => env('JWT_TTL', 3600),
+    'ttl' => (int) env('JWT_TTL', 3600),
 
     /*
     |--------------------------------------------------------------------------
@@ -117,7 +117,7 @@ return [
     |
     */
 
-    'refresh_ttl' => env('JWT_REFRESH_TTL', 3600 * 24 * 14),
+    'refresh_ttl' => (int) env('JWT_REFRESH_TTL', 3600 * 24 * 14),
 
     /*
     |--------------------------------------------------------------------------
@@ -204,7 +204,7 @@ return [
     |
     */
 
-    'leeway' => env('JWT_LEEWAY', 0),
+    'leeway' => (int) env('JWT_LEEWAY', 0),
 
     /*
     |--------------------------------------------------------------------------
@@ -231,7 +231,7 @@ return [
     |
     */
 
-    'blacklist_grace_period' => env('JWT_BLACKLIST_GRACE_PERIOD', 0),
+    'blacklist_grace_period' => (int) env('JWT_BLACKLIST_GRACE_PERIOD', 0),
 
     /*
     |--------------------------------------------------------------------------
